+++ example/lib/homepage/network_tools/bili_favorite_video/bili_favorite_video_logic.dart
@@ -4,7 +4,8 @@ import 'package:redstonex/redstonex.dart';
 
 import 'bili_favorite_video_state.dart';
 
-class BiliFavoriteVideoLogic extends PagingController2<BiliFavoriteVideo, BiliFavoriteVideoPagingState> {
+class BiliFavoriteVideoLogic extends EventPagingController<BiliFavoriteVideo, BiliFavoriteVideoPagingState,
+    BiliFavoriteVideoListRefreshableEvent<BiliFavoriteVideo>> {
   final BiliFavoriteVideoState state = BiliFavoriteVideoState();
   final BiliService _biliService = XDepends().on();
 
@@ -19,8 +20,9 @@ class BiliFavoriteVideoLogic extends PagingController2<BiliFavoriteVideo, BiliFa
   }
 
   @override
-  PagingParams getPagingParams() {
-    int minId = state.biliFavoriteVideoPagingState.data.isEmpty ? -1 : state.biliFavoriteVideoPagingState.data.last.value.id!;
+  PagingParams customPagingParams() {
+    int minId =
+        state.biliFavoriteVideoPagingState.data.isEmpty ? -1 : state.biliFavoriteVideoPagingState.data.last.value.id!;
     return PagingParams.create(
         pageIndex: state.biliFavoriteVideoPagingState.nextIndex,
         pageSize: state.biliFavoriteVideoPagingState.pageSize,