+++ lib/networks/exception/api_exception.dart
@@ -52,7 +52,7 @@ class ApiException implements Exception {
             default:
               return ApiException(errCode, error.response?.statusMessage ?? '未知错误');
           }
-        } on Exception catch (e) {
+        } on Exception {
           return ApiException(_apiCommonExceptionCode, unknownException);
         }
       default: