+++ lib/paging/paging_params.g.dart
@@ -10,9 +10,7 @@ PagingParams _$PagingParamsFromJson(Map<String, dynamic> json) => PagingParams()
   ..currentIndex = json['currentIndex'] as int
   ..extra = json['extra'] as Map<String, dynamic>?
   ..model = json['model'] as Map<String, dynamic>
-  ..order = json['order'] as String?
   ..size = json['size'] as int
-  ..sort = json['sort'] as String?
   ..total = json['total'] as int?;
 
 Map<String, dynamic> _$PagingParamsToJson(PagingParams instance) =>
@@ -20,8 +18,6 @@ Map<String, dynamic> _$PagingParamsToJson(PagingParams instance) =>
       'currentIndex': instance.currentIndex,
       'extra': instance.extra,
       'model': instance.model,
-      'order': instance.order,
       'size': instance.size,
-      'sort': instance.sort,
       'total': instance.total,
     };